Update a shadow when acquiring mechanical state
The state of the lock could be changed in ways other than the API,
which can lead to a mismatch between the real state and the device shadow.
For instance, `toggle` is looking at a shadow, this mismatch may
prevent it from controlling a key correctly.
This commit is aims to be a little more aggresive update a shadow.